refactor(c-api): Move `JanetArray` to use `size_t` on `count` and `capacity fields`

+++ src/include/janet.h
@@ -965,8 +965,8 @@ struct JanetStackFrame {
 /* A dynamic array type. */
 struct JanetArray {
     JanetGCObject gc;
-    int32_t count;
-    int32_t capacity;
+    size_t count;
+    size_t capacity;
     Janet *data;
 };
